MMc
I have been playing around with the MMc code by the creator of Salabim along with the old sim.py file Josh created. It needs some work for the statistics part, we also need to add a way to change the number of lanes, the inter-arrival time and the server time on the command line. I will be working on it some more.